src/requirements.txt
onnxruntime==1.11.0
onnx==1.11.0
onnx-simplifier==0.3.7
numpy==1.22.3
opencv-python==4.5.5.64
